split out the nav and breadcrumb items from the trust test specs
